+++ package.json
@@ -15,6 +15,7 @@
     "@tailwindcss/vite": "^4.1.4",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
+    "react-hook-form": "^7.56.0",
     "tailwindcss": "^4.1.4"
   },
   "devDependencies": {
