+++ src/lib/types/types.ts
@@ -1,0 +1,24 @@
+export type InputProps = {
+  label?: string;
+  placeholder?: string;
+  value?: string;
+  onChange?: (
+    e: React.ChangeEvent<HTMLInputElement | HTMLSelectElement>
+  ) => void;
+  className?: string;
+  name?: string;
+  type?: string;
+};
+
+export type CardProps = {
+  children: React.ReactNode;
+  className?: string;
+};
+
+export type ButtonProps = {
+  label: string;
+  onClick?: () => void;
+  disabled?: boolean;
+  className?: string;
+  icon?: React.ReactNode;
+};
