+++ src/lib/ui/Card.tsx
@@ -1,0 +1,7 @@
+import { CardProps } from '../types/types';
+
+const Card = ({ children, className }: CardProps) => {
+  return <div className={className}>{children}</div>;
+};
+
+export default Card;
