--- src/types/button.ts
@@ -1,7 +1,0 @@
-export type ButtonProps = {
-  label: string;
-  onClick?: () => void;
-  disabled?: boolean;
-  className?: string;
-  icon?: React.ReactNode;
-};
