--- src/types/card.ts
@@ -1,4 +1,0 @@
-export type CardProps = {
-  children: React.ReactNode;
-  className?: string;
-};
