--- src/types/input.ts
@@ -1,9 +1,0 @@
-export type InputProps = {
-  label?: string;
-  type?: string;
-  placeholder?: string;
-  value?: string;
-  onChange?: (e: React.ChangeEvent<HTMLInputElement>) => void;
-  className?: string;
-  name?: string;
-};
